backend/requirements.txt
flask
flask-cors
scikit-learn
pandas
numpy
joblib
flask-jwt-extended
flask-bcrypt
flask-sqlalchemy
Pillow
gunicorn
requests
python-dotenv
tensorflow-cpu
numpy

